+++ PlannerCRM/Shared/DTOs/EmployeeDto/Forms/EmployeeSalaryDTO.cs
@@ -4,7 +4,7 @@ public class EmployeeSalaryDto
 {
     public int Id { get; set; }
     public int EmployeeId { get; set; }
-    public float Salary { get; set; }
+    public decimal Salary { get; set; }
     public DateTime StartDate { get; set; }
     public DateTime FinishDate { get; set; }
 }
