+++ PlannerCRM/Shared/DTOs/EmployeeDto/Views/EmployeeViewDTO.cs
@@ -11,7 +11,7 @@ public class EmployeeViewDto
     public string LastName { get; set; }
     public string FullName { get; set; }
     public string Role { get; set; }
-    public float HourlyRate { get; set; }
+    public decimal HourlyRate { get; set; }
     public DateTime BirthDay { get; set; }
     public DateTime StartDate { get; set; }
     public List<EmployeeSalaryDto> EmployeeSalaries { get; set; }
